backend/requirements.txt
fastapi
uvicorn
requests
google-genai
beautifulsoup4
python-dotenv
apscheduler
lxml
